+++ test/features/auth/presentation/providers/auth_provider_test.mocks.dart
@@ -9,24 +9,22 @@ import 'package:dartz/dartz.dart' as _i3;
 import 'package:mockito/mockito.dart' as _i1;
 import 'package:tkit/core/errors/failure.dart' as _i6;
 import 'package:tkit/features/auth/data/datasources/token_local_datasource.dart'
-    as _i13;
+    as _i12;
 import 'package:tkit/features/auth/data/models/device_code_response.dart'
-    as _i16;
-import 'package:tkit/features/auth/data/models/twitch_token_model.dart' as _i14;
-import 'package:tkit/features/auth/data/models/twitch_user_model.dart' as _i15;
-import 'package:tkit/features/auth/domain/entities/twitch_token.dart' as _i11;
-import 'package:tkit/features/auth/domain/entities/twitch_user.dart' as _i7;
+    as _i15;
+import 'package:tkit/features/auth/data/models/twitch_token_model.dart' as _i13;
+import 'package:tkit/features/auth/data/models/twitch_user_model.dart' as _i14;
+import 'package:tkit/features/auth/domain/entities/twitch_token.dart' as _i9;
+import 'package:tkit/features/auth/domain/entities/twitch_user.dart' as _i11;
 import 'package:tkit/features/auth/domain/repositories/i_auth_repository.dart'
     as _i2;
-import 'package:tkit/features/auth/domain/usecases/authenticate_usecase.dart'
-    as _i4;
 import 'package:tkit/features/auth/domain/usecases/check_auth_status_usecase.dart'
-    as _i9;
+    as _i7;
 import 'package:tkit/features/auth/domain/usecases/get_current_user_usecase.dart'
-    as _i12;
-import 'package:tkit/features/auth/domain/usecases/logout_usecase.dart' as _i8;
-import 'package:tkit/features/auth/domain/usecases/refresh_token_usecase.dart'
     as _i10;
+import 'package:tkit/features/auth/domain/usecases/logout_usecase.dart' as _i4;
+import 'package:tkit/features/auth/domain/usecases/refresh_token_usecase.dart'
+    as _i8;
 
 // ignore_for_file: type=lint
 // ignore_for_file: avoid_redundant_argument_values
@@ -53,45 +51,10 @@ class _FakeEither_1<L, R> extends _i1.SmartFake implements _i3.Either<L, R> {
     : super(parent, parentInvocation);
 }
 
-/// A class which mocks [AuthenticateUseCase].
-///
-/// See the documentation for Mockito's code generation for more information.
-class MockAuthenticateUseCase extends _i1.Mock
-    implements _i4.AuthenticateUseCase {
-  MockAuthenticateUseCase() {
-    _i1.throwOnMissingStub(this);
-  }
-
-  @override
-  _i2.IAuthRepository get repository =>
-      (super.noSuchMethod(
-            Invocation.getter(#repository),
-            returnValue: _FakeIAuthRepository_0(
-              this,
-              Invocation.getter(#repository),
-            ),
-          )
-          as _i2.IAuthRepository);
-
-  @override
-  _i5.Future<_i3.Either<_i6.Failure, _i7.TwitchUser>> call() =>
-      (super.noSuchMethod(
-            Invocation.method(#call, []),
-            returnValue:
-                _i5.Future<_i3.Either<_i6.Failure, _i7.TwitchUser>>.value(
-                  _FakeEither_1<_i6.Failure, _i7.TwitchUser>(
-                    this,
-                    Invocation.method(#call, []),
-                  ),
-                ),
-          )
-          as _i5.Future<_i3.Either<_i6.Failure, _i7.TwitchUser>>);
-}
-
 /// A class which mocks [LogoutUseCase].
 ///
 /// See the documentation for Mockito's code generation for more information.
-class MockLogoutUseCase extends _i1.Mock implements _i8.LogoutUseCase {
+class MockLogoutUseCase extends _i1.Mock implements _i4.LogoutUseCase {
   MockLogoutUseCase() {
     _i1.throwOnMissingStub(this);
   }
@@ -125,7 +88,7 @@ class MockLogoutUseCase extends _i1.Mock implements _i8.LogoutUseCase {
 ///
 /// See the documentation for Mockito's code generation for more information.
 class MockCheckAuthStatusUseCase extends _i1.Mock
-    implements _i9.CheckAuthStatusUseCase {
+    implements _i7.CheckAuthStatusUseCase {
   MockCheckAuthStatusUseCase() {
     _i1.throwOnMissingStub(this);
   }
@@ -159,7 +122,7 @@ class MockCheckAuthStatusUseCase extends _i1.Mock
 ///
 /// See the documentation for Mockito's code generation for more information.
 class MockRefreshTokenUseCase extends _i1.Mock
-    implements _i10.RefreshTokenUseCase {
+    implements _i8.RefreshTokenUseCase {
   MockRefreshTokenUseCase() {
     _i1.throwOnMissingStub(this);
   }
@@ -176,25 +139,25 @@ class MockRefreshTokenUseCase extends _i1.Mock
           as _i2.IAuthRepository);
 
   @override
-  _i5.Future<_i3.Either<_i6.Failure, _i11.TwitchToken>> call() =>
+  _i5.Future<_i3.Either<_i6.Failure, _i9.TwitchToken>> call() =>
       (super.noSuchMethod(
             Invocation.method(#call, []),
             returnValue:
-                _i5.Future<_i3.Either<_i6.Failure, _i11.TwitchToken>>.value(
-                  _FakeEither_1<_i6.Failure, _i11.TwitchToken>(
+                _i5.Future<_i3.Either<_i6.Failure, _i9.TwitchToken>>.value(
+                  _FakeEither_1<_i6.Failure, _i9.TwitchToken>(
                     this,
                     Invocation.method(#call, []),
                   ),
                 ),
           )
-          as _i5.Future<_i3.Either<_i6.Failure, _i11.TwitchToken>>);
+          as _i5.Future<_i3.Either<_i6.Failure, _i9.TwitchToken>>);
 }
 
 /// A class which mocks [GetCurrentUserUseCase].
 ///
 /// See the documentation for Mockito's code generation for more information.
 class MockGetCurrentUserUseCase extends _i1.Mock
-    implements _i12.GetCurrentUserUseCase {
+    implements _i10.GetCurrentUserUseCase {
   MockGetCurrentUserUseCase() {
     _i1.throwOnMissingStub(this);
   }
@@ -211,31 +174,31 @@ class MockGetCurrentUserUseCase extends _i1.Mock
           as _i2.IAuthRepository);
 
   @override
-  _i5.Future<_i3.Either<_i6.Failure, _i7.TwitchUser?>> call() =>
+  _i5.Future<_i3.Either<_i6.Failure, _i11.TwitchUser?>> call() =>
       (super.noSuchMethod(
             Invocation.method(#call, []),
             returnValue:
-                _i5.Future<_i3.Either<_i6.Failure, _i7.TwitchUser?>>.value(
-                  _FakeEither_1<_i6.Failure, _i7.TwitchUser?>(
+                _i5.Future<_i3.Either<_i6.Failure, _i11.TwitchUser?>>.value(
+                  _FakeEither_1<_i6.Failure, _i11.TwitchUser?>(
                     this,
                     Invocation.method(#call, []),
                   ),
                 ),
           )
-          as _i5.Future<_i3.Either<_i6.Failure, _i7.TwitchUser?>>);
+          as _i5.Future<_i3.Either<_i6.Failure, _i11.TwitchUser?>>);
 }
 
 /// A class which mocks [TokenLocalDataSource].
 ///
 /// See the documentation for Mockito's code generation for more information.
 class MockTokenLocalDataSource extends _i1.Mock
-    implements _i13.TokenLocalDataSource {
+    implements _i12.TokenLocalDataSource {
   MockTokenLocalDataSource() {
     _i1.throwOnMissingStub(this);
   }
 
   @override
-  _i5.Future<void> saveToken(_i14.TwitchTokenModel? token) =>
+  _i5.Future<void> saveToken(_i13.TwitchTokenModel? token) =>
       (super.noSuchMethod(
             Invocation.method(#saveToken, [token]),
             returnValue: _i5.Future<void>.value(),
@@ -244,12 +207,12 @@ class MockTokenLocalDataSource extends _i1.Mock
           as _i5.Future<void>);
 
   @override
-  _i5.Future<_i14.TwitchTokenModel?> getToken() =>
+  _i5.Future<_i13.TwitchTokenModel?> getToken() =>
       (super.noSuchMethod(
             Invocation.method(#getToken, []),
-            returnValue: _i5.Future<_i14.TwitchTokenModel?>.value(),
+            returnValue: _i5.Future<_i13.TwitchTokenModel?>.value(),
           )
-          as _i5.Future<_i14.TwitchTokenModel?>);
+          as _i5.Future<_i13.TwitchTokenModel?>);
 
   @override
   _i5.Future<void> deleteToken() =>
@@ -269,7 +232,7 @@ class MockTokenLocalDataSource extends _i1.Mock
           as _i5.Future<bool>);
 
   @override
-  _i5.Future<void> saveUser(_i15.TwitchUserModel? user) =>
+  _i5.Future<void> saveUser(_i14.TwitchUserModel? user) =>
       (super.noSuchMethod(
             Invocation.method(#saveUser, [user]),
             returnValue: _i5.Future<void>.value(),
@@ -278,12 +241,12 @@ class MockTokenLocalDataSource extends _i1.Mock
           as _i5.Future<void>);
 
   @override
-  _i5.Future<_i15.TwitchUserModel?> getUser() =>
+  _i5.Future<_i14.TwitchUserModel?> getUser() =>
       (super.noSuchMethod(
             Invocation.method(#getUser, []),
-            returnValue: _i5.Future<_i15.TwitchUserModel?>.value(),
+            returnValue: _i5.Future<_i14.TwitchUserModel?>.value(),
           )
-          as _i5.Future<_i15.TwitchUserModel?>);
+          as _i5.Future<_i14.TwitchUserModel?>);
 
   @override
   _i5.Future<void> deleteUser() =>
@@ -313,44 +276,30 @@ class MockIAuthRepository extends _i1.Mock implements _i2.IAuthRepository {
   }
 
   @override
-  _i5.Future<_i3.Either<_i6.Failure, _i7.TwitchUser>> authenticate() =>
-      (super.noSuchMethod(
-            Invocation.method(#authenticate, []),
-            returnValue:
-                _i5.Future<_i3.Either<_i6.Failure, _i7.TwitchUser>>.value(
-                  _FakeEither_1<_i6.Failure, _i7.TwitchUser>(
-                    this,
-                    Invocation.method(#authenticate, []),
-                  ),
-                ),
-          )
-          as _i5.Future<_i3.Either<_i6.Failure, _i7.TwitchUser>>);
-
-  @override
-  _i5.Future<_i3.Either<_i6.Failure, _i16.DeviceCodeResponse>>
+  _i5.Future<_i3.Either<_i6.Failure, _i15.DeviceCodeResponse>>
   initiateDeviceCodeAuth() =>
       (super.noSuchMethod(
             Invocation.method(#initiateDeviceCodeAuth, []),
             returnValue:
                 _i5.Future<
-                  _i3.Either<_i6.Failure, _i16.DeviceCodeResponse>
+                  _i3.Either<_i6.Failure, _i15.DeviceCodeResponse>
                 >.value(
-                  _FakeEither_1<_i6.Failure, _i16.DeviceCodeResponse>(
+                  _FakeEither_1<_i6.Failure, _i15.DeviceCodeResponse>(
                     this,
                     Invocation.method(#initiateDeviceCodeAuth, []),
                   ),
                 ),
           )
-          as _i5.Future<_i3.Either<_i6.Failure, _i16.DeviceCodeResponse>>);
+          as _i5.Future<_i3.Either<_i6.Failure, _i15.DeviceCodeResponse>>);
 
   @override
-  _i5.Future<_i3.Either<_i6.Failure, _i7.TwitchUser>>
+  _i5.Future<_i3.Either<_i6.Failure, _i11.TwitchUser>>
   authenticateWithDeviceCode(String? deviceCode) =>
       (super.noSuchMethod(
             Invocation.method(#authenticateWithDeviceCode, [deviceCode]),
             returnValue:
-                _i5.Future<_i3.Either<_i6.Failure, _i7.TwitchUser>>.value(
-                  _FakeEither_1<_i6.Failure, _i7.TwitchUser>(
+                _i5.Future<_i3.Either<_i6.Failure, _i11.TwitchUser>>.value(
+                  _FakeEither_1<_i6.Failure, _i11.TwitchUser>(
                     this,
                     Invocation.method(#authenticateWithDeviceCode, [
                       deviceCode,
@@ -358,7 +307,7 @@ class MockIAuthRepository extends _i1.Mock implements _i2.IAuthRepository {
                   ),
                 ),
           )
-          as _i5.Future<_i3.Either<_i6.Failure, _i7.TwitchUser>>);
+          as _i5.Future<_i3.Either<_i6.Failure, _i11.TwitchUser>>);
 
   @override
   _i5.Future<_i3.Either<_i6.Failure, void>> logout() =>
@@ -374,18 +323,18 @@ class MockIAuthRepository extends _i1.Mock implements _i2.IAuthRepository {
           as _i5.Future<_i3.Either<_i6.Failure, void>>);
 
   @override
-  _i5.Future<_i3.Either<_i6.Failure, _i11.TwitchToken>> refreshToken() =>
+  _i5.Future<_i3.Either<_i6.Failure, _i9.TwitchToken>> refreshToken() =>
       (super.noSuchMethod(
             Invocation.method(#refreshToken, []),
             returnValue:
-                _i5.Future<_i3.Either<_i6.Failure, _i11.TwitchToken>>.value(
-                  _FakeEither_1<_i6.Failure, _i11.TwitchToken>(
+                _i5.Future<_i3.Either<_i6.Failure, _i9.TwitchToken>>.value(
+                  _FakeEither_1<_i6.Failure, _i9.TwitchToken>(
                     this,
                     Invocation.method(#refreshToken, []),
                   ),
                 ),
           )
-          as _i5.Future<_i3.Either<_i6.Failure, _i11.TwitchToken>>);
+          as _i5.Future<_i3.Either<_i6.Failure, _i9.TwitchToken>>);
 
   @override
   _i5.Future<_i3.Either<_i6.Failure, bool>> isAuthenticated() =>
@@ -401,30 +350,30 @@ class MockIAuthRepository extends _i1.Mock implements _i2.IAuthRepository {
           as _i5.Future<_i3.Either<_i6.Failure, bool>>);
 
   @override
-  _i5.Future<_i3.Either<_i6.Failure, _i11.TwitchToken?>> getCurrentToken() =>
+  _i5.Future<_i3.Either<_i6.Failure, _i9.TwitchToken?>> getCurrentToken() =>
       (super.noSuchMethod(
             Invocation.method(#getCurrentToken, []),
             returnValue:
-                _i5.Future<_i3.Either<_i6.Failure, _i11.TwitchToken?>>.value(
-                  _FakeEither_1<_i6.Failure, _i11.TwitchToken?>(
+                _i5.Future<_i3.Either<_i6.Failure, _i9.TwitchToken?>>.value(
+                  _FakeEither_1<_i6.Failure, _i9.TwitchToken?>(
                     this,
                     Invocation.method(#getCurrentToken, []),
                   ),
                 ),
           )
-          as _i5.Future<_i3.Either<_i6.Failure, _i11.TwitchToken?>>);
+          as _i5.Future<_i3.Either<_i6.Failure, _i9.TwitchToken?>>);
 
   @override
-  _i5.Future<_i3.Either<_i6.Failure, _i7.TwitchUser?>> getCurrentUser() =>
+  _i5.Future<_i3.Either<_i6.Failure, _i11.TwitchUser?>> getCurrentUser() =>
       (super.noSuchMethod(
             Invocation.method(#getCurrentUser, []),
             returnValue:
-                _i5.Future<_i3.Either<_i6.Failure, _i7.TwitchUser?>>.value(
-                  _FakeEither_1<_i6.Failure, _i7.TwitchUser?>(
+                _i5.Future<_i3.Either<_i6.Failure, _i11.TwitchUser?>>.value(
+                  _FakeEither_1<_i6.Failure, _i11.TwitchUser?>(
                     this,
                     Invocation.method(#getCurrentUser, []),
                   ),
                 ),
           )
-          as _i5.Future<_i3.Either<_i6.Failure, _i7.TwitchUser?>>);
+          as _i5.Future<_i3.Either<_i6.Failure, _i11.TwitchUser?>>);
 }